requirements.txt
psycopg2-binary==2.9.11
Django==5.2.8
pillow==12.0.0